Moved Transformer to its own file
